@@ -1,7 +1,12 @@
 import React from 'react'
+import PostsGallery from './PostsGallery'
 
-export default function PostsContainer() {
+export default function PostsContainer({ posts }) {
+
   return (
-    <div>PostsContainer</div>
+    <div>
+        PostsContainer
+        <PostsGallery posts={posts} />
+    </div>
   )
 }
