@@ -1,6 +1,13 @@
-import React from 'react'
+import React, { useState } from 'react'
+import { useNavigate } from 'react-router-dom'
 
+const emptyForm = {
+  name:''
+
+}
 export default function SignUp() {
+  
+  
   return (
     <div>SignUp</div>
   )
