@@ -1,7 +1,18 @@
 import React from 'react'
 
-export default function UserContainer() {
+export default function UserProfile({ user }) {
+  const { name, instagram, image, city, frame, bio, age, professional } = user
   return (
-    <div>UserContainer</div>
+    <div>
+      UserContainer
+      {name}
+      {instagram}
+      {image}
+      {city}
+      {frame}
+      {bio}
+      {age}
+      {professional}
+    </div>
   )
 }
